backend/requirements.txt
# Core FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP and API clients
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
aiofiles==23.2.1

# Database and caching
motor==3.3.2
pymongo==4.6.1

# Image and media processing
cloudinary==1.36.0
Pillow==10.1.0

# Web scraping
beautifulsoup4==4.12.2
selenium==4.21.0
webdriver-manager==4.0.1
lxml==4.9.3
fake-useragent==1.4.0
undetected-chromedriver==3.5.4
cloudscraper==1.2.71
scrapy==2.11.0

# Data processing
pandas==2.1.4
numpy==1.24.3
plotly==5.17.0

# Text processing and sentiment analysis
vaderSentiment==3.3.2
textblob==0.17.1
nltk==3.8.1

# Reddit API
praw==7.7.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Production deployment
gunicorn==21.2.0

# Enhanced Analysis Dependencies
scrapy==2.11.0
scrapy-user-agents==0.1.1
scrapy-rotating-proxies==0.6.2
scrapy-splash==0.8.0
crochet==2.0.0
twisted==22.10.0
praw==7.7.1
prawcore==2.3.0
textblob==0.17.1
nltk==3.8.1
spacy==3.7.2
transformers==4.36.0
torch==2.1.0
asyncio-pool==0.6.0
concurrent-futures==3.1.1
celery==5.3.4
redis==5.0.1